autoecole/requirements.txt
anyio==4.10.0
asgiref==3.8.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.6
django-cors-headers==4.4.0
django-environ==0.12.0
django-extensions==3.2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
firebase_admin==7.1.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
msgpack==1.1.1
numpy==2.1.0
pandas==2.2.2
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydotplus==2.0.2
PyJWT==2.10.1
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2022.1
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.2
typing_extensions==4.12.2
tzdata==2022.7
urllib3==2.5.0
